+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-* feat: Add instrumentation helpers for Android Compose
+* refactor: Move touch auto-instrumentation to separate package
+* feat: Add instrumentation package for Android Compose
 
 ## v0.0.3-alpha